+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# `2.0.1`
+
+- Fix bug in base64url decoding: now accept base64url data with or without padding bytes
+
 # `2.0.0`
 
 - Fix platform minimum versions being too low for iOS, tvOS, and watchOS